Adding props to latest news
