+++ src/components/LatestNews/LatestNews.module.css
@@ -1,19 +1,15 @@
-.swiper-container {
-  display: flex;
-  margin: 0 auto;
-  align-items: center;
-  gap: clamp(1rem, 5rem, 10rem);
-  max-width: 125rem;
+.section-container {
   background-color: var(--bg-dark-blue);
   padding: 8rem;
+  display: flex;
+  flex-direction: column;
+  margin: 0 auto;
+  max-width: 125rem;
+  gap: 6rem;
 }
 
-.arrow-icon {
-  color: var(--bg-pink-brown);
-  background: rgba(0, 0, 0, 0);
-}
-
-.swiper-nav-prev,
-.swiper-nav-next {
-  cursor: pointer;
+.text-container {
+  display: flex;
+  justify-content: space-between;
+  padding: 0 clamp(0rem, 10rem, 15rem);
 }
