+++ src/components/LatestNews/latestNewsSlider/sliderData.ts
@@ -1,0 +1,72 @@
+const sliderData = [
+  {
+    id: 1,
+    imageSrc:
+      "https://dev.sagafarmann.com/wp/wp-content/uploads/2023/03/DJI_0858.jpg",
+    heading: "heading",
+    alt: "Captain",
+    text: "Lorem ipsum dolor sit amet consectetur adipisicing elit. Maxime molliti molestiae quas vel sint commodi repudiandae consequuntur voluptatum laborum numquam blanditiis harum quisquam eius sed odit",
+  },
+
+  {
+    id: 2,
+    imageSrc:
+      "https://dev.sagafarmann.com/wp/wp-content/uploads/2023/03/DJI_0858.jpg",
+    heading: "heading",
+    alt: "Captain",
+    text: "Lorem ipsum dolor sit amet consectetur adipisicing elit. Maxime molliti molestiae quas vel sint commodi repudiandae consequuntur voluptatum laborum numquam blanditiis harum quisquam eius sed odit",
+  },
+  {
+    id: 3,
+    imageSrc:
+      "https://dev.sagafarmann.com/wp/wp-content/uploads/2023/03/DJI_0858.jpg",
+    heading: "heading",
+    alt: "Captain",
+    text: "Lorem ipsum dolor sit amet consectetur adipisicing elit. Maxime molliti molestiae quas vel sint commodi repudiandae consequuntur voluptatum laborum numquam blanditiis harum quisquam eius sed odit",
+  },
+
+  {
+    id: 4,
+    imageSrc:
+      "https://dev.sagafarmann.com/wp/wp-content/uploads/2023/03/DJI_0858.jpg",
+    heading: "heading",
+    alt: "Captain",
+    text: "Lorem ipsum dolor sit amet consectetur adipisicing elit. Maxime molliti molestiae quas vel sint commodi repudiandae consequuntur voluptatum laborum numquam blanditiis harum quisquam eius sed odit",
+  },
+  {
+    id: 5,
+    imageSrc:
+      "https://dev.sagafarmann.com/wp/wp-content/uploads/2023/03/DJI_0858.jpg",
+    heading: "heading",
+    alt: "Captain",
+    text: "Lorem ipsum dolor sit amet consectetur adipisicing elit. Maxime molliti molestiae quas vel sint commodi repudiandae consequuntur voluptatum laborum numquam blanditiis harum quisquam eius sed odit",
+  },
+
+  {
+    id: 6,
+    imageSrc:
+      "https://dev.sagafarmann.com/wp/wp-content/uploads/2023/03/DJI_0858.jpg",
+    heading: "heading",
+    alt: "Captain",
+    text: "Lorem ipsum dolor sit amet consectetur adipisicing elit. Maxime molliti molestiae quas vel sint commodi repudiandae consequuntur voluptatum laborum numquam blanditiis harum quisquam eius sed odit",
+  },
+  {
+    id: 7,
+    imageSrc:
+      "https://dev.sagafarmann.com/wp/wp-content/uploads/2023/03/DJI_0858.jpg",
+    heading: "heading",
+    alt: "Captain",
+    text: "Lorem ipsum dolor sit amet consectetur adipisicing elit. Maxime molliti molestiae quas vel sint commodi repudiandae consequuntur voluptatum laborum numquam blanditiis harum quisquam eius sed odit",
+  },
+
+  {
+    id: 8,
+    imageSrc:
+      "https://dev.sagafarmann.com/wp/wp-content/uploads/2023/03/DJI_0858.jpg",
+    heading: "heading",
+    alt: "Captain",
+    text: "Lorem ipsum dolor sit amet consectetur adipisicing elit. Maxime molliti molestiae quas vel sint commodi repudiandae consequuntur voluptatum laborum numquam blanditiis harum quisquam eius sed odit",
+  },
+];
+
+export default sliderData;
