+++ src/components/cards/sliderCard/SliderCard.module.css
@@ -1,5 +1,8 @@
 .card {
+  display: flex;
+  flex-direction: column;
   height: 46rem;
+  max-width: 36rem;
   background-color: var(--very-dark-blue);
   border-radius: 0.8rem;
   overflow: hidden;
