+++ src/styles/globals.css
@@ -87,12 +87,12 @@ time,
 mark,
 audio,
 video {
-	margin: 0;
-	padding: 0;
-	border: 0;
-	font-size: 100%;
-	font: inherit;
-	vertical-align: baseline;
+  margin: 0;
+  padding: 0;
+  border: 0;
+  font-size: 100%;
+  font: inherit;
+  vertical-align: baseline;
 }
 /* HTML5 display-role reset for older browsers */
 article,
@@ -106,65 +106,65 @@ hgroup,
 menu,
 nav,
 section {
-	display: block;
+  display: block;
 }
 body {
-	line-height: 1;
+  line-height: 1;
 }
 ol,
 ul {
-	list-style: none;
+  list-style: none;
 }
 blockquote,
 q {
-	quotes: none;
+  quotes: none;
 }
 blockquote:before,
 blockquote:after,
 q:before,
 q:after {
-	content: "";
-	content: none;
+  content: "";
+  content: none;
 }
 table {
-	border-collapse: collapse;
-	border-spacing: 0;
+  border-collapse: collapse;
+  border-spacing: 0;
 }
 
 html {
-	font-size: 62.5%;
+  font-size: 62.5%;
 }
 
 body {
-	font-size: 1.6rem;
+  font-size: 1.6rem;
 }
 @font-face {
-	font-family: "Comfortaa";
-	src: url("../components/fonts/comfortaa/Comfortaa-Regular.eot");
-	src: url("../components/fonts/comfortaa/Comfortaa-Regular.eot?#iefix")
-			format("embedded-opentype"),
-		url("../components/fonts/comfortaa/Comfortaa-Regular.woff2") format("woff2"),
-		url("../components/fonts/comfortaa/Comfortaa-Regular.woff") format("woff"),
-		url("../components/fonts/comfortaa/Comfortaa-Regular.ttf")
-			format("truetype"),
-		url("../components/fonts/comfortaa/Comfortaa-Regular.svg#Comfortaa-Regular")
-			format("svg");
-	font-weight: normal;
-	font-style: normal;
-	font-display: swap;
+  font-family: "Comfortaa";
+  src: url("../components/fonts/comfortaa/Comfortaa-Regular.eot");
+  src: url("../components/fonts/comfortaa/Comfortaa-Regular.eot?#iefix")
+      format("embedded-opentype"),
+    url("../components/fonts/comfortaa/Comfortaa-Regular.woff2") format("woff2"),
+    url("../components/fonts/comfortaa/Comfortaa-Regular.woff") format("woff"),
+    url("../components/fonts/comfortaa/Comfortaa-Regular.ttf")
+      format("truetype"),
+    url("../components/fonts/comfortaa/Comfortaa-Regular.svg#Comfortaa-Regular")
+      format("svg");
+  font-weight: normal;
+  font-style: normal;
+  font-display: swap;
 }
 @font-face {
-	font-family: "Norse";
-	src: url("../components/fonts/norse/Norse.eot");
-	src: url("../components/fonts/norse/Norse.eot?#iefix")
-			format("embedded-opentype"),
-		url("../components/fonts/norse/Norse.woff2") format("woff2"),
-		url("../components/fonts/norse/Norse.woff") format("woff"),
-		url("../components/fonts/norse/Norse.ttf") format("truetype"),
-		url("../components/fonts/norse/Norse.svg#Norse") format("svg");
-	font-weight: normal;
-	font-style: normal;
-	font-display: swap;
+  font-family: "Norse";
+  src: url("../components/fonts/norse/Norse.eot");
+  src: url("../components/fonts/norse/Norse.eot?#iefix")
+      format("embedded-opentype"),
+    url("../components/fonts/norse/Norse.woff2") format("woff2"),
+    url("../components/fonts/norse/Norse.woff") format("woff"),
+    url("../components/fonts/norse/Norse.ttf") format("truetype"),
+    url("../components/fonts/norse/Norse.svg#Norse") format("svg");
+  font-weight: normal;
+  font-style: normal;
+  font-display: swap;
 }
 
 .swiper {
